+++ src/main/kotlin/com/example/cmc_be/notification/convertor/NotificationConvertor.kt
@@ -1,8 +1,8 @@
-package com.example.cmc_be.notice.convertor
+package com.example.cmc_be.notification.convertor
 
 import com.example.cmc_be.common.annotation.Convertor
 import com.example.cmc_be.domain.notification.entity.Notification
-import com.example.cmc_be.notice.dto.NotificationRes
+import com.example.cmc_be.notification.dto.NotificationRes
 
 @Convertor
 class NotificationConvertor {
